+++ lib/enumerize.rb
@@ -14,6 +14,13 @@ module Enumerize
 
   autoload :ActiveRecordSupport, 'enumerize/activerecord'
   autoload :SequelSupport, 'enumerize/sequel'
+  autoload :MongoidSupport,      'enumerize/mongoid'
+
+  module Scope
+    autoload :ActiveRecord, 'enumerize/scope/activerecord'
+    autoload :Sequel, 'enumerize/scope/sequel'
+    autoload :Mongoid,      'enumerize/scope/mongoid'
+  end
 
   def self.included(base)
     ActiveSupport::Deprecation.warn '`include Enumerize` was deprecated. Please use `extend Enumerize`.', caller
@@ -23,8 +30,21 @@ module Enumerize
   def self.extended(base)
     base.send :include, Enumerize::Base
     base.extend Enumerize::Predicates
-    base.extend Enumerize::ActiveRecordSupport
-    base.extend Enumerize::SequelSupport
+
+    if defined?(::ActiveRecord::Base)
+      base.extend Enumerize::ActiveRecordSupport
+      base.extend Enumerize::Scope::ActiveRecord
+    end
+
+    if defined?(::Mongoid::Document)
+      base.extend Enumerize::MongoidSupport
+      base.extend Enumerize::Scope::Mongoid
+    end
+
+    if defined?(::Sequel::Model)
+      base.extend Enumerize::SequelSupport
+      base.extend Enumerize::Scope::Sequel
+    end
 
     if defined?(::RailsAdmin)
       require 'enumerize/integrations/rails_admin'