+++ lib/enumerize/value.rb
@@ -17,18 +17,18 @@ module Enumerize
       I18n.t(i18n_keys[0], :default => i18n_keys[1..-1])
     end
 
+    def ==(other)
+      super(other.to_s) || value == other
+    end
+
     def encode_with(coder)
       coder.represent_object(self.class.superclass, @value)
     end
 
     private
 
-    def define_query_method(value)
-      singleton_class.class_eval <<-RUBY, __FILE__, __LINE__ + 1
-        def #{value}?
-          #{value == self}
-        end
-      RUBY
+    def predicate_call(value)
+      value == self
     end
 
     def i18n_keys
@@ -36,7 +36,7 @@ module Enumerize
         i18n_keys = i18n_scopes
         i18n_keys << [:"enumerize.defaults.#{@attr.name}.#{self}"]
         i18n_keys << [:"enumerize.#{@attr.name}.#{self}"]
-        i18n_keys << self.humanize # humanize value if there are no translations
+        i18n_keys << self.underscore.humanize # humanize value if there are no translations
         i18n_keys.flatten
       end
     end