+++ test/formtastic_test.rb
@@ -1,16 +1,7 @@
 require 'test_helper'
-require 'rails'
 
-module Formtastic
-  module Helpers
-    module InputHelper
-      remove_method :input_class
-      def input_class(as)
-        input_class_with_const_defined(as)
-      end
-    end
-  end
-end
+Formtastic::FormBuilder.action_class_finder = Formtastic::ActionClassFinder
+Formtastic::FormBuilder.input_class_finder  = Formtastic::InputClassFinder
 
 class FormtasticSpec < MiniTest::Spec
   include ViewTestHelper