+++ test/set_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require 'yaml'
 
 describe Enumerize::Set do
   let(:klass) do
@@ -103,6 +104,12 @@ describe Enumerize::Set do
     end
   end
 
+  describe '#texts' do
+    it 'returns array of text values' do
+      set.texts.must_equal ['A']
+    end
+  end
+
   describe '#join' do
     it 'joins values' do
       set << :b
@@ -145,6 +152,13 @@ describe Enumerize::Set do
 
     it 'does not respond to a method for not existing value' do
       set.wont_respond_to :some_method?
+    end
+  end
+
+  describe 'serialization' do
+    it 'is serialized to yaml as array' do
+      set << :a
+      assert_equal YAML.dump(%w(a)), YAML.dump(set)
     end
   end
 end