+++ test/test_helper.rb
@@ -2,11 +2,32 @@ require 'minitest/autorun'
 require 'minitest/spec'
 require 'active_support/core_ext/kernel/reporting'
 require 'active_model'
-
-$VERBOSE=true
+require 'rails'
+begin
+  require 'mongoid'
+rescue LoadError
+end
 
 module RailsAdmin
 end
+
+require 'simple_form'
+SimpleForm.setup {}
+
+require 'formtastic'
+
+module EnumerizeTest
+  class Application < Rails::Application
+    config.active_support.deprecation = :stderr
+    config.active_support.test_order = :random
+    config.eager_load = false
+    config.secret_key_base = 'secret'
+  end
+end
+
+EnumerizeTest::Application.initialize!
+
+$VERBOSE=true
 
 require 'enumerize'